Decouple AccountValidator from Account (and its specs)
Using `validate!` couples the validator heavily with the Account model. This is because we're forced to handle exceptions without providing any alternatives. This forces us to use specific classes to catch exceptions.

By making this change (skipping loading an account number) we can then use the `errors` array from the Validator to then choose if having an validation error (say, a duplicate) is something exceptional or something worth stopping execution.